git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@828 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/registrartable.c
@@ -746,7 +746,8 @@ int registrarTableGetRegistrar(struct RegistrarTable*   registrarTable,
                               registrarTable->OutstandingConnects--;
                            }
                            LOG_VERBOSE3
-                           fprintf(stdlog, "Failed to establish registrar connection, outstanding=%u\n", registrarTable->OutstandingConnects);
+                           fprintf(stdlog, "Failed to establish registrar connection, outstanding=%u\n",
+                                   (unsigned int)registrarTable->OutstandingConnects);
                            LOG_END
                         }
                   }
